backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx[http2]==0.27.2
websockets==12.0
PyJWT==2.8.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg>=0.30.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
respx==0.21.1

# Linting
black==24.10.0
flake8==7.1.1
isort==5.13.2
